requirements.txt
opencv-python==4.5.5.64
tensorflow==2.2
keras==2.4.3
imutils==0.5.4
underthesea==1.3.4
scikit-learn==0.24.2
protobuf==3.20.1
keras_applications
facenet_pytorch
fasttext
google.cloud

